requirements.txt
uvicorn
fastapi
python-dotenv


















#google-generativeai
gTTs
python-multipart
